requirements.txt
pyyaml==6.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
azure-identity==1.15.0
azure-storage-blob==12.19.0